dell-backend-final/requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.5
django-celery-results==2.2.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
joblib==1.1.0
kombu==5.2.3
mysql-connector-python==8.0.28
# mysqlclient==2.1.0
numpy==1.22.1
packaging==21.3
Pillow==8.3.1
prompt-toolkit==3.0.26
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.2
requests==2.26.0
# scikit-learn==1.0.2
# scipy==1.7.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.0.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
gunicorn==20.1.0
